requirements.txt
streamlit
cohere
requests
transformers
python-dotenv
